func_react_stream_serverless/requirements.txt
cohere
anthropic==0.30.1
beautifulsoup4==4.12.3
fastapi==0.111.0
fastapi-cli==0.0.5
Flask==3.0.1
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flask-restful-swagger-2==0.35
greenlet==3.0.3
groq==0.9.0
htbuilder==0.6.2
html2text==2024.2.26
httpcore==1.0.2
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
instructor==1.4.0
ipykernel==6.28.0
ipython==8.19.0
json5==0.9.14
json_repair==0.8.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
langchain==0.2.7
langchain-anthropic==0.1.19
langchain-community==0.2.7
langchain-core==0.2.12
langchain-experimental==0.0.62
langchain-google-vertexai==1.0.6
langchain-groq==0.1.6
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langgraph==0.1.5
langserve==0.2.2
langsmith==0.1.83
launchpadlib==1.10.16
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.12
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
openai==1.42.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
pycparser==2.21
pydantic==2.8.0
pydantic_core==2.20.0
pyparsing==2.4.7
PyPDF2==3.0.1
quantile-python==1.1
referencing==0.30.2
regex==2023.8.8
requests==2.32.3
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rootpath==0.1.1
rouge-score==0.1.2
rpds-py==0.9.2
sentence-transformers==2.3.1
sgmllib3k==1.0.0
shapely==2.0.4
shellingham==1.5.4
shortuuid==1.0.11
simple-websocket==1.0.0
six==1.16.0
slowapi==0.1.9
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlitedict==2.1.0
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-chat==0.1.1
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
svgwrite==1.4.3
sympy==1.12
syncer==2.0.3
tabledata==1.3.3
tabulate==0.9.0
tavily-python==0.3.3
tcolorpy==0.1.4
tenacity==8.5.0
termcolor==2.4.0
terminado==0.18.0
texttable==1.7.0
threadpoolctl==3.2.0
tiktoken==0.7.0
tinycss2==1.2.1
together==1.2.12
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tox==4.12.1
tqdm==4.66.5
tqdm-multiprocess==0.0.11
traitlets==5.14.1
transformers==4.41.2
Tree==0.2.4
typepy==1.3.2
typer==0.12.5
types-python-dateutil==2.8.19.14
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.0.4
uvicorn==0.30.1
uvloop==0.19.0
weaviate-client==3.26.2
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.8
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.1
wikipedia==1.4.0
Wikipedia-API==0.6.0

